producer/requirements.txt
kafka-python==2.0.2
faker==24.0.0
python-dotenv==1.0.1
google-cloud-pubsub==2.21.1